:truck: renamed helpers.h -> Helpers.h
Just to be consistent, since we're using Helpers for the namespace
